requirements.txt
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.26.0b1
azure-search-documents==11.7.0b2
azure-storage-blob==12.28.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
msal==1.35.0b1
msal-extensions==1.3.1
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
